Added visual for `nearToons`
This red orb is a visual representation of how the far the V.P. is capable of using his directed attack. Any toon who is within this radius has a chance of being selected for a targeted gear attack. If no toons are in this radius, the V.P. will always use his Jump attack.