Wrote minor clarification on trackExp
Given the formula in the guide, one may think that (for example), a Dropless toon using a Drop SOS will have a trackExp of -10, since they don't have any gags in the track ((0 - 1) * 10). However, the game actually treats all toons as having the level 1 gag in every track by default. The tracks however, are not unlocked until the appropriate tasks are taken throughout the game.